# for-agent-dev/05-backend-api-openapi.yaml
openapi: 3.0.3
info:
  title: Whats Awesome Backend API
  version: 0.2.0
  description: |
    Whats Awesome 后端标准接口文档。

    当前实现为 Node Express 本地骨架，接口契约对齐后续华为云 APIG + FunctionGraph + GaussDB。
    REST 接口分为 Portal 人类用户 API、Admin 人类管理 API、Admin-Agent 管理侧 Agent API。
    用户面开发者 Agent 不调用 REST 点亮接口，后续走 Portal MCP 服务；MCP 契约见 `for-agent-dev/07-portal-mcp-contract.md`。
servers:
  - url: http://localhost:8000
    description: Local backend
...
paths:
...
  /api/admin/quests/{id}/review:
    post:
      tags: [Admin]
      summary: 管理员审核点亮申请
      x-implementation-status: implemented
      security:
        - AdminSession: []
      parameters:
        - $ref: '#/components/parameters/Id'
        - name: x-admin-id
          in: header
          required: true
          schema: { type: string }
          description: |
            迭代二本地白名单: admin-founder / admin-secretary。
            正式版本由 GitCode OAuth + RBAC 注入管理员身份。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewQuestRequest'
      responses:
        '200':
          description: 审核完成。通过时可能返回新发放勋章。
          content:
            application/json:
              schema:
                type: object
                required: [quest, granted_badges]
                properties:
                  quest:
                    $ref: '#/components/schemas/Quest'
                  granted_badges:
                    type: array
                    items:
                      $ref: '#/components/schemas/Badge'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 申请已经审核过
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/admin/skill-candidates:
    get:
      tags: [Admin]
      summary: 查询技能候选
      description: 管理端读取技能 Agent 通过 /api/admin-agent/skill-candidates 提交的候选。
      x-implementation-status: implemented
      security:
        - AdminSession: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CandidateStatus'
      responses:
        '200':
          description: 技能候选列表
          content:
            application/json:
              schema:
                type: object
                required: [total, items]
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SkillCandidate'
  /api/admin/skill-candidates/{id}/approve:
    post:
      tags: [Admin]
      summary: 审核通过技能候选
      x-implementation-status: implemented
      security:
        - AdminSession: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateReviewRequest'
      responses:
        '200':
          description: 候选已发布或合并到技能库
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateReviewResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/skill-candidates/{id}/reject:
    post:
      tags: [Admin]
      summary: 驳回技能候选
      x-implementation-status: implemented
      security:
        - AdminSession: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateReviewRequest'
      responses:
        '200':
          description: 候选已驳回
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateReviewResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/news-candidates:
    get:
      tags: [Admin]
      summary: 查询资讯候选
      description: 管理端读取技能 Agent 通过 /api/admin-agent/news-candidates 提交的资讯候选。
      x-implementation-status: implemented
      security:
        - AdminSession: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/CandidateStatus'
      responses:
        '200':
          description: 资讯候选列表
          content:
            application/json:
              schema:
                type: object
                required: [total, items]
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/NewsCandidate'
  /api/admin/news-candidates/{id}/publish:
    post:
      tags: [Admin]
      summary: 发布资讯候选
      x-implementation-status: implemented
      security:
        - AdminSession: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateReviewRequest'
      responses:
        '200':
          description: 资讯已发布
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateReviewResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/admin/news-candidates/{id}/reject:
    post:
      tags: [Admin]
      summary: 驳回资讯候选
      x-implementation-status: implemented
      security:
        - AdminSession: []
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateReviewRequest'
      responses:
        '200':
          description: 资讯候选已驳回
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateReviewResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin-agent/skill-candidates:
    post:
      tags: [AdminAgent]
      summary: 技能 Agent 提交技能候选
      description: 只写入 pending_review，不直接发布到用户面。
      x-implementation-status: implemented
      security:
        - AdminAgentId: []
          AdminAgentSignature: []
          AdminAgentTimestamp: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SkillCandidateSubmission'
      responses:
        '201':
          description: 技能候选已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillCandidate'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin-agent/news-candidates:
    post:
      tags: [AdminAgent]
      summary: 技能 Agent 提交资讯候选
      description: 只写入 pending_review，管理端审核发布后用户面可见。
      x-implementation-status: implemented
      security:
        - AdminAgentId: []
          AdminAgentSignature: []
          AdminAgentTimestamp: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsCandidateSubmission'
      responses:
        '201':
          description: 资讯候选已提交
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsCandidate'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    UserSession:
      type: apiKey
      in: cookie
      name: wa_session
      description: GitCode OAuth 登录后的用户会话。
    AdminSession:
      type: apiKey
      in: cookie
      name: wa_admin_session
      description: GitCode OAuth + RBAC 后的管理员会话；本地临时用 x-admin-id 模拟。
    AdminAgentId:
      type: apiKey
      in: header
      name: x-admin-agent-id
      description: 管理侧内部 Agent 客户端 ID。
    AdminAgentSignature:
      type: apiKey
      in: header
      name: x-admin-agent-signature
      description: 管理侧 Agent 请求签名，密钥存 DEW/CSMS，不写入仓库。
    AdminAgentTimestamp:
      type: apiKey
      in: header
      name: x-admin-agent-timestamp
      description: 签名时间戳，用于防重放。
  parameters:
...
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
    IdempotencyKey:
      name: x-idempotency-key
      in: header
      required: true
      schema: { type: string }
      description: 写接口幂等键，避免定时任务重试造成重复候选。
  responses:
    ValidationError:
      description: 请求参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 没有权限
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
...
    CandidateStatus:
      type: string
      enum: [pending_review, approved, rejected, merged]
...
    SkillCandidate:
      allOf:
        - $ref: '#/components/schemas/SkillCandidateSubmission'
        - type: object
          required: [id, status, created_at]
          properties:
            id: { type: string }
            status:
              $ref: '#/components/schemas/CandidateStatus'
            review_note: { type: string }
            created_at: { type: string, format: date-time }
...
    CandidateReviewRequest:
      type: object
      properties:
        review_note: { type: string }
        publish_slug: { type: string }
    CandidateReviewResponse:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        status:
          $ref: '#/components/schemas/CandidateStatus'
        published_slug: { type: string }
        review_note: { type: string }
...
    Badge:
      type: object
      required: [key, name, title, icon, description, granted_at]
      properties:
        key: { type: string }
        name: { type: string }
        title: { type: string }
        icon: { type: string }
        description: { type: string }
        granted_at: { type: string, format: date-time }
...
    TargetType:
      type: string
      enum: [skill, case]
    JudgeStatus:
      type: string
      enum: [pending, approved, rejected]
    Quest:
      type: object
      required: [id, player_id, target_type, target_slug, method, evidence, judge_status]
      properties:
        id: { type: string }
        player_id: { type: string }
        target_type:
          $ref: '#/components/schemas/TargetType'
        target_slug: { type: string }
        method:
          type: string
          enum: [manual_upload, mcp_auto, admin_grant]
        evidence:
          $ref: '#/components/schemas/Evidence'
        judge_status:
          $ref: '#/components/schemas/JudgeStatus'
        judged_by: { type: string }
        judge_note: { type: string }
        lit_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ReviewQuestRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approved, rejected]
        judge_note: { type: string }
...